agri-analysis-server/requirements.txt
cffi==1.14.6
click==8.0.1
cmdstanpy==0.9.68
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
ephem==4.0.0.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
hijri-converter==2.1.3
holidays==0.11.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
plotly==5.1.0
prophet==1.0.1
py4j==0.10.9
pyarrow==4.0.1
pycparser==2.20
PyJWT==2.1.0
PyMeeus==0.5.11
PyMySQL==1.0.2
pyparsing==2.4.7
pyspark==3.0.3
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.7.0
setuptools-git==1.2
six==1.16.0
SQLAlchemy==1.4.22
SQLAlchemy-serializer==1.4.1
tenacity==8.0.1
threadpoolctl==2.2.0
tqdm==4.61.2
ujson==4.0.2
Werkzeug==2.0.1
